+++ packages/rest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ts-fetcher/rest
 
+## 1.3.0
+
+### Minor Changes
+
+- Added `force` option for request cache configuration <br>
+  If this option provided, it will override old cache and execute http request
+
 ## 1.2.1
 
 ### Patch Changes